Another Kolor Theme Update
Modify Foreground Color and JavaScript Colors in the Dark Theme

Changed the foreground color of text in the Dark theme from a light grey
to white in order to make content stand out better in HTML markup.
Changed the JavaScript function call colors to add more color
variability. When testing with a local Gulp configuration file, there
was too much gold in the color scheme.

* Changed global foreground color from #969696 to #FFFFFF (line 31)
* Changed global selection color from #3F352A to #3C3C3C (line 43)
* Changed global selection border color from #3F352A to #3C3C3C (line
* 45)
* Redirect to the stored location after successfully logging in the user
* Changed comment color from #525252 to #3D3D3D (line 58)
* Changed variable color from #FCD660 to #abe9fb (line 157)
* Changed function call color from #FCD660 to #BDEC35 (line 264)